consumer/requirements.txt
Flask
kafka-python
redis